move new function to _WF
